fully working bot
