--- .env
@@ -1,5 +1,0 @@
-#!/bin/bash
-export CONSUMER_KEY=""
-export CONSUMER_SECRET=""
-export ACCESS_TOKEN=""
-export ACCESS_TOKEN_SECRET=""
