+++ .gitignore
@@ -1,2 +1,3 @@
 .vscode
 .env
+__pycache__
