+++ create_api.py
@@ -5,6 +5,7 @@ import os
 
 logger = logging.getLogger()
 
+
 def create_api():
     logger.info("Loading environment variables")
     load_dotenv()
@@ -12,7 +13,7 @@ def create_api():
     consumer_secret = os.getenv("CONSUMER_SECRET")
     access_token = os.getenv("ACCESS_TOKEN")
     access_token_secret = os.getenv("ACCESS_TOKEN_SECRET")
- 
+
     logger.info("Initialising Twitter API")
     auth = tweepy.OAuthHandler(consumer_key, consumer_secret)
     auth.set_access_token(access_token, access_token_secret)
